fix: 댓글 삭제 시 작성자를 받도록 수정

+++ src/main/java/com/ndgl/spotfinder/domain/comment/service/PostCommentService.java
@@ -83,8 +83,12 @@ public class PostCommentService {
 	}
 
 	@Transactional
-	public void delete(Long id, Long commentId) {
+	public void delete(Long id, Long commentId, String email) {
+		User author = userRepository.findByEmail(email)
+			.orElseThrow(ErrorCode.USER_NOT_FOUND::throwServiceException);
+
 		PostComment comment = findCommentAndVerifyPost(commentId, id);
+		comment.checkAuthorCanDelete(author);
 		postCommentRepository.delete(comment);
 	}
